fix(quant): newcalc probe 按 fp32 中间量预算 TileM, 修 16KB tile 不可编码
kBudgetElems 原按 sizeof(__half) 预算, 令 kTilemMax=128, 但 data pass 的
fp32 tile_f 物理字节 = TileM*BlockSize*4 = 16KB, 超出 TSize 仅编码 <=8KB 2 的幂
的上限 (pto_tile.hpp TilesizeCode -> __tilesize_unknown), 大 M 编不过。改按
sizeof(float) 预算 -> kTilemMax=64 (tile_f=8KB 合法)。M=512/N=256 现可编,
gfsim 35236 cyc / 0.269 cyc-per-elem。探针 cpp 参数化为 -DPM/-DPN 便于扫尺寸。

+++ benchmark/one-level-arch/kernels/quant/dynamic_mx_quant/probe_dynamic_mx_quant_tail_ocp_fp8_newcalc.hpp
@@ -56,7 +56,10 @@ void probe_dynamic_mx_quant_tail_ocp_fp8_newcalc(__half *x, __fp8_e4m3 *y, uint8
     constexpr uint16_t RECIP_XOR_NOT   = 0xFFFF; // 按位取反 (int16: -1)
     constexpr uint16_t RECIP_COMPL_SUB = 0x80FF; // 0xFFFF - 0x7F00 (int16 补码减法, 低16位不变)
 
-    constexpr int kBudgetElems = 8192 / static_cast<int>(sizeof(__half)); // 4096
+    // 按最大中间量 (data pass 的 fp32 tile_f) 预算, 而非 half: tile_f 物理字节 =
+    // TileM*BlockSize*4, 而 TSize 只编码 <=8KB 的 2 的幂, 故须 TileM*BlockSize*4<=8192。
+    // 用 sizeof(__half) 预算会漏掉这翻倍, 令 TileM 撑到 128 -> tile_f=16KB -> 不可编码。
+    constexpr int kBudgetElems = 8192 / static_cast<int>(sizeof(float)); // 2048 (fp32-budgeted)
     constexpr int kMinElems    = 512  / static_cast<int>(sizeof(__half)); // 256
     constexpr int kTilemMax    = kBudgetElems / BlockSize;
     constexpr int kTilemMin    = (kMinElems + BlockSize - 1) / BlockSize;
